Refactor Banking.scala to Firm/Household quality level
BankStatus ADT (Active/Failed) replaces 3 flat Boolean/Int fields.
Named constants for all Basel III weights and calibration numbers.
Opaque types: nplRatio/car/lcr/nsfr → Ratio, hqla/asf/rsf → PLN,
method params PLN/Rate where appropriate.
Named result types: PerBankAmounts, FailureCheckResult, BailInResult,
ResolutionResult replace raw tuples.
BankState fields have no defaults — all constructors explicit.
Inline field comments matching Firm.State/Household.State pattern.
Test mkBank helpers reduce constructor boilerplate.

31 files changed, 1246 tests pass.

+++ src/test/scala/sfc/engine/CorporateBondSpec.scala
@@ -197,5 +197,5 @@ class CorporateBondSpec extends AnyFlatSpec with Matchers:
       corpBondHoldings = PLN(400.0),
     )
     // RWA = 1000 + 400 * 0.5 = 1200; CAR = 200 / 1200 = 0.1667
-    bank.car shouldBe (200.0 / 1200.0 +- 0.001)
+    bank.car.toDouble shouldBe (200.0 / 1200.0 +- 0.001)
   }
